Fix ResizeBilinear GPU index overflow
ResizeBilinear calculated input offsets with a 32-bit batch index. The offset could overflow and read outside the tensor when the input had more than INT32_MAX elements.

I changed the batch index to int64 so the input offset does not overflow.

+++ tensorflow/core/kernels/image/resize_bilinear_op_gpu.cu.cc
@@ -296,7 +296,7 @@ __global__ void LegacyResizeBilinearKernel(
     const int x = idx % out_width;
     idx /= out_width;
     const int y = idx % out_height;
-    const int b = idx / out_height;
+    const int64_t b = idx / out_height;
 
     const float in_y = y * height_scale;
     const int top_y_index = floorf(in_y);
